+++ cmd/zombiezen-sqlite-migrate/testdata/TestProcess/ExecScriptFS/want.go
@@ -1,0 +1,10 @@
+// Copyright 2021 Ross Light
+// SPDX-License-Identifier: ISC
+
+package main
+
+import "zombiezen.com/go/sqlite/sqlitex"
+
+func main() {
+	_ = sqlitex.ExecuteScriptFS
+}