+++ cmd/zombiezen-sqlite-migrate/testdata/skeleton/zombiezen.com/go/sqlite/stubs.go
@@ -1,0 +1,11 @@
+// Copyright 2021 Ross Light
+// SPDX-License-Identifier: ISC
+
+// Test stubs for zombiezen.com/go/sqlite.
+package sqlite
+
+import (
+	"os"
+)
+
+type Conn struct{}