+++ sqlitex/exec_test.go
@@ -15,7 +15,7 @@
 //
 // SPDX-License-Identifier: ISC
 
-package sqlitex_test
+package sqlitex
 
 import (
 	"fmt"
@@ -23,7 +23,6 @@ import (
 	"testing"
 
 	"zombiezen.com/go/sqlite"
-	"zombiezen.com/go/sqlite/sqlitex"
 )
 
 func TestExec(t *testing.T) {
@@ -33,13 +32,13 @@ func TestExec(t *testing.T) {
 	}
 	defer conn.Close()
 
-	if err := sqlitex.ExecTransient(conn, "CREATE TABLE t (a TEXT, b INTEGER);", nil); err != nil {
+	if err := ExecTransient(conn, "CREATE TABLE t (a TEXT, b INTEGER);", nil); err != nil {
 		t.Fatal(err)
 	}
-	if err := sqlitex.Exec(conn, "INSERT INTO t (a, b) VALUES (?, ?);", nil, "a1", 1); err != nil {
+	if err := Exec(conn, "INSERT INTO t (a, b) VALUES (?, ?);", nil, "a1", 1); err != nil {
 		t.Error(err)
 	}
-	if err := sqlitex.Exec(conn, "INSERT INTO t (a, b) VALUES (?, ?);", nil, "a2", 2); err != nil {
+	if err := Exec(conn, "INSERT INTO t (a, b) VALUES (?, ?);", nil, "a2", 2); err != nil {
 		t.Error(err)
 	}
 
@@ -50,7 +49,7 @@ func TestExec(t *testing.T) {
 		b = append(b, stmt.ColumnInt64(1))
 		return nil
 	}
-	if err := sqlitex.ExecTransient(conn, "SELECT a, b FROM t;", fn); err != nil {
+	if err := ExecTransient(conn, "SELECT a, b FROM t;", fn); err != nil {
 		t.Fatal(err)
 	}
 	if want := []string{"a1", "a2"}; !reflect.DeepEqual(a, want) {
@@ -68,7 +67,7 @@ func TestExecErr(t *testing.T) {
 	}
 	defer conn.Close()
 
-	err = sqlitex.Exec(conn, "INVALID SQL STMT", nil)
+	err = Exec(conn, "INVALID SQL STMT", nil)
 	if err == nil {
 		t.Error("invalid SQL did not return an error code")
 	}
@@ -76,16 +75,16 @@ func TestExecErr(t *testing.T) {
 		t.Errorf("INVALID err code=%s, want %s", got, want)
 	}
 
-	if err := sqlitex.Exec(conn, "CREATE TABLE t (c1, c2);", nil); err != nil {
+	if err := Exec(conn, "CREATE TABLE t (c1, c2);", nil); err != nil {
 		t.Error(err)
 	}
-	if err := sqlitex.Exec(conn, "INSERT INTO t (c1, c2) VALUES (?, ?);", nil, 1, 1); err != nil {
+	if err := Exec(conn, "INSERT INTO t (c1, c2) VALUES (?, ?);", nil, 1, 1); err != nil {
 		t.Error(err)
 	}
-	if err := sqlitex.Exec(conn, "INSERT INTO t (c1, c2) VALUES (?, ?);", nil, 2, 2); err != nil {
+	if err := Exec(conn, "INSERT INTO t (c1, c2) VALUES (?, ?);", nil, 2, 2); err != nil {
 		t.Error(err)
 	}
-	err = sqlitex.Exec(conn, "INSERT INTO t (c1, c2) VALUES (?, ?);", nil, 1, 1, 1)
+	err = Exec(conn, "INSERT INTO t (c1, c2) VALUES (?, ?);", nil, 1, 1, 1)
 	if got, want := sqlite.ErrCode(err), sqlite.ResultRange; got != want {
 		t.Errorf("INSERT err code=%s, want %s", got, want)
 	}
@@ -96,13 +95,87 @@ func TestExecErr(t *testing.T) {
 		calls++
 		return customErr
 	}
-	err = sqlitex.Exec(conn, "SELECT c1 FROM t;", fn)
+	err = Exec(conn, "SELECT c1 FROM t;", fn)
 	if err != customErr {
 		t.Errorf("SELECT want err=customErr, got: %v", err)
 	}
 	if calls != 1 {
 		t.Errorf("SELECT want truncated callback calls, got calls=%d", calls)
 	}
+}
+
+func TestExecArgsErrors(t *testing.T) {
+	conn, err := sqlite.OpenConn(":memory:", 0)
+	if err != nil {
+		t.Fatal(err)
+	}
+	defer conn.Close()
+
+	t.Run("TooManyPositional", func(t *testing.T) {
+		err := Exec(conn, `SELECT ?;`, nil, 1, 2)
+		t.Log(err)
+		if got, want := sqlite.ErrCode(err), sqlite.ResultRange; got != want {
+			t.Errorf("code = %v; want %v", got, want)
+		}
+	})
+
+	t.Run("Missing", func(t *testing.T) {
+		// Compatibility: crawshaw.io/sqlite does not check for this condition.
+		err := Exec(conn, `SELECT ?;`, nil)
+		t.Log(err)
+		if got, want := sqlite.ErrCode(err), sqlite.ResultOK; got != want {
+			t.Errorf("code = %v; want %v", got, want)
+		}
+	})
+}
+
+func TestExecuteArgsErrors(t *testing.T) {
+	conn, err := sqlite.OpenConn(":memory:", 0)
+	if err != nil {
+		t.Fatal(err)
+	}
+	defer conn.Close()
+
+	t.Run("TooManyPositional", func(t *testing.T) {
+		err := Execute(conn, `SELECT ?;`, &ExecOptions{
+			Args: []interface{}{1, 2},
+		})
+		t.Log(err)
+		if got, want := sqlite.ErrCode(err), sqlite.ResultRange; got != want {
+			t.Errorf("code = %v; want %v", got, want)
+		}
+	})
+
+	t.Run("ExtraNamed", func(t *testing.T) {
+		err := Execute(conn, `SELECT :foo;`, &ExecOptions{
+			Named: map[string]interface{}{
+				":foo": 42,
+				":bar": "hi",
+			},
+		})
+		t.Log(err)
+		if got, want := sqlite.ErrCode(err), sqlite.ResultRange; got != want {
+			t.Errorf("code = %v; want %v", got, want)
+		}
+	})
+
+	t.Run("Missing", func(t *testing.T) {
+		err := Execute(conn, `SELECT ?;`, &ExecOptions{
+			Args: []interface{}{},
+		})
+		t.Log(err)
+		if got, want := sqlite.ErrCode(err), sqlite.ResultError; got != want {
+			t.Errorf("code = %v; want %v", got, want)
+		}
+	})
+
+	t.Run("MissingNamed", func(t *testing.T) {
+		err := Execute(conn, `SELECT :foo;`, nil)
+		t.Log(err)
+		if got, want := sqlite.ErrCode(err), sqlite.ResultError; got != want {
+			t.Errorf("code = %v; want %v", got, want)
+		}
+	})
 }
 
 func TestExecScript(t *testing.T) {
@@ -118,7 +191,7 @@ INSERT INTO t (a, b) VALUES ('a1', 1);
 INSERT INTO t (a, b) VALUES ('a2', 2);
 `
 
-	if err := sqlitex.ExecScript(conn, script); err != nil {
+	if err := ExecScript(conn, script); err != nil {
 		t.Error(err)
 	}
 
@@ -127,11 +200,162 @@ INSERT INTO t (a, b) VALUES ('a2', 2);
 		sum = stmt.ColumnInt(0)
 		return nil
 	}
-	if err := sqlitex.Exec(conn, "SELECT sum(b) FROM t;", fn); err != nil {
+	if err := Exec(conn, "SELECT sum(b) FROM t;", fn); err != nil {
 		t.Fatal(err)
 	}
 
 	if sum != 3 {
 		t.Errorf("sum=%d, want 3", sum)
+	}
+}
+
+func TestExecuteScript(t *testing.T) {
+	conn, err := sqlite.OpenConn(":memory:", 0)
+	if err != nil {
+		t.Fatal(err)
+	}
+	defer conn.Close()
+
+	script := `
+CREATE TABLE t (a TEXT, b INTEGER);
+INSERT INTO t (a, b) VALUES ('a1', :a1);
+INSERT INTO t (a, b) VALUES ('a2', :a2);
+`
+
+	err = ExecuteScript(conn, script, &ExecOptions{
+		Named: map[string]interface{}{
+			":a1": 1,
+			":a2": 2,
+		},
+	})
+	if err != nil {
+		t.Error(err)
+	}
+
+	sum := 0
+	fn := func(stmt *sqlite.Stmt) error {
+		sum = stmt.ColumnInt(0)
+		return nil
+	}
+	if err := Exec(conn, "SELECT sum(b) FROM t;", fn); err != nil {
+		t.Fatal(err)
+	}
+
+	if sum != 3 {
+		t.Errorf("sum=%d, want 3", sum)
+	}
+
+	t.Run("ExtraNamed", func(t *testing.T) {
+		conn, err := sqlite.OpenConn(":memory:", 0)
+		if err != nil {
+			t.Fatal(err)
+		}
+		defer conn.Close()
+
+		script := `
+CREATE TABLE t (a TEXT, b INTEGER);
+INSERT INTO t (a, b) VALUES ('a1', :a1);
+INSERT INTO t (a, b) VALUES ('a2', :a2);
+`
+
+		err = ExecuteScript(conn, script, &ExecOptions{
+			Named: map[string]interface{}{
+				":a1": 1,
+				":a2": 2,
+				":a3": 3,
+			},
+		})
+		t.Log(err)
+		if got, want := sqlite.ErrCode(err), sqlite.ResultRange; got != want {
+			t.Errorf("code = %v; want %v", got, want)
+		}
+	})
+
+	t.Run("MissingNamed", func(t *testing.T) {
+		conn, err := sqlite.OpenConn(":memory:", 0)
+		if err != nil {
+			t.Fatal(err)
+		}
+		defer conn.Close()
+
+		script := `
+CREATE TABLE t (a TEXT, b INTEGER);
+INSERT INTO t (a, b) VALUES ('a1', :a1);
+INSERT INTO t (a, b) VALUES ('a2', :a2);
+`
+
+		err = ExecuteScript(conn, script, &ExecOptions{
+			Named: map[string]interface{}{
+				":a1": 1,
+			},
+		})
+		t.Log(err)
+		if got, want := sqlite.ErrCode(err), sqlite.ResultError; got != want {
+			t.Errorf("code = %v; want %v", got, want)
+		}
+	})
+}
+
+func TestBitsetHasAll(t *testing.T) {
+	tests := []struct {
+		bs   bitset
+		n    int
+		want bool
+	}{
+		{
+			bs:   bitset{},
+			n:    0,
+			want: true,
+		},
+		{
+			bs:   bitset{0},
+			n:    1,
+			want: false,
+		},
+		{
+			bs:   bitset{0x0000000000000001},
+			n:    1,
+			want: true,
+		},
+		{
+			bs:   bitset{0x8000000000000001},
+			n:    1,
+			want: true,
+		},
+		{
+			bs:   bitset{0x0000000000000001},
+			n:    2,
+			want: false,
+		},
+		{
+			bs:   bitset{0xffffffffffffffff},
+			n:    64,
+			want: true,
+		},
+		{
+			bs:   bitset{0xffffffffffffffff},
+			n:    65,
+			want: false,
+		},
+		{
+			bs:   bitset{0xffffffffffffffff, 0x0000000000000000},
+			n:    65,
+			want: false,
+		},
+		{
+			bs:   bitset{0xffffffffffffffff, 0x0000000000000001},
+			n:    65,
+			want: true,
+		},
+		{
+			bs:   bitset{0x7fffffffffffffff, 0x0000000000000001},
+			n:    65,
+			want: false,
+		},
+	}
+	for _, test := range tests {
+		if got := test.bs.hasAll(test.n); got != test.want {
+			t.Errorf("%v.hasAll(%d) = %t; want %t", test.bs, test.n, got, test.want)
+		}
 	}
 }